requirements.txt
tensorflow
numpy
requests
Pillow
scikit-learn
instagrapi
rich
flask
gunicorn
